remove constants
