+++ src/main/java/frc/robot/RobotContainer.java
@@ -19,7 +19,8 @@ import frc.robot.hid.DemonCommandXboxController;
 import frc.robot.hid.OperatorStick;
 import frc.robot.subsystems.Climber;
 import frc.robot.subsystems.algae.collector.AlgaeCollector;
-import frc.robot.subsystems.cameras.CameraConfig;
+import frc.robot.subsystems.cameras.CameraConfig.PhotonVisionConfig;
+import frc.robot.subsystems.cameras.CameraConfig.LimelightConfig;
 import frc.robot.subsystems.cameras.DemonLimelightCamera;
 import frc.robot.subsystems.cameras.DemonPhotonCamera;
 import frc.robot.subsystems.coraller.Coraller;
@@ -84,9 +85,9 @@ public class RobotContainer {
 
   // Subsystems - Cameras
   public static final DemonPhotonCamera kLowerFrontPhotonCamera = new DemonPhotonCamera(
-      CameraConfig.PhotonvisionSetup.kLowerFrontCameraName, CameraConfig.PhotonvisionSetup.kRobotToLowerFrontCamera);
+      PhotonVisionConfig.kLowerFrontCameraName, PhotonVisionConfig.kRobotToLowerFrontCamera);
   public static final DemonLimelightCamera kRearLimelightCamera = new DemonLimelightCamera(
-    CameraConfig.LimelightConfig.kRearCameraName, CameraConfig.LimelightConfig.kPoseAlgorithm, kSwerveDrive::getPose, kNavx::getRate);
+    LimelightConfig.kRearCameraName, LimelightConfig.kPoseAlgorithm, kSwerveDrive::getPose, kNavx::getRate);
 
   /**
    * The container for the robot. Contains subsystems, operator interface devices,
