+++ src/main/java/frc/robot/subsystems/algae/collector/AlgaeCollectorConfig.java
@@ -1,0 +1,11 @@
+package frc.robot.subsystems.algae.collector;
+
+public class AlgaeCollectorConfig {
+  public static final class AlgaeWristConfig {
+    public static final double kAlgaeP = 0.0;
+    public static final double kAlgaeI = 0.0;
+    public static final double kAlgaeD = 0.0;
+    /** Gravity term in volts. */
+    public static final double kAlageKG = 0.0;
+  }
+}
